feat(Sync-VS-Async): add 리액티브 형식의 콜백을 이용해 해결하는 방법 추가
