+++ src/main/java/me/withbeth/mij/concurrency/UsingCallback.java
@@ -1,0 +1,23 @@
+package me.withbeth.mij.concurrency;
+
+import java.util.function.IntConsumer;
+
+public class UsingCallback {
+
+    public static void main(String[] args) {
+        int x = 1;
+
+        PairResult.Builder resultBuilder = new PairResult.Builder();
+
+        ExternalAPIs.callbackF(x, value -> {
+            resultBuilder.left(value);
+            System.out.println(resultBuilder);
+        });
+
+        ExternalAPIs.callbackG(x, value -> {
+            resultBuilder.right(value);
+            System.out.println(resultBuilder);
+        });
+
+    }
+}
