+++ src/main/java/me/withbeth/mij/concurrency/UsingThreadManually.java
@@ -21,30 +21,4 @@ public class UsingThreadManually {
         System.out.println("result = " + result);
     }
 
-    record PairResult(int left, int right) {
-        static class Builder {
-            private int left;
-            private int right;
-            Builder left(int val) {
-                left = val;
-                return this;
-            }
-            Builder right(int val) {
-                right = val;
-                return this;
-            }
-            PairResult build() {
-                return new PairResult(left, right);
-            }
-        }
-
-        @Override
-        public String toString() {
-            return "{" +
-                    "left=" + left +
-                    ", right=" + right +
-                    '}';
-        }
-    }
-
 }
